Corretti errori

+++ service/database/db-ban.go
@@ -1,8 +1,8 @@
 package database
 
 //BanUser() banna l'utente banned
-func (db *appdbimpl) BanUser(banner integer, banned integer) error {
-	_, err := db.c.Exec("INSERT INTO ban (banner,banned) VALUES (?, ?)", banner, banned)
+func (db *appdbimpl) BanUser(banner User, banned User) error {
+	_, err := db.c.Exec("INSERT INTO ban (banner,banned) VALUES (?, ?)", banner.uId, banned.uId)
 	if err != nil {
 		// Errore nell'esecuzione della query
 		return err
@@ -12,8 +12,8 @@ func (db *appdbimpl) BanUser(banner integer, banned integer) error {
 }
 
 //UnbanUser() sbanna l'utente banned
-func (db *appdbimpl) UnbanUser(banner integer, banned integer) error {
-	_, err := db.c.Exec("DELETE FROM ban WHERE banner = ? AND banned = ?", banner, banned)
+func (db *appdbimpl) UnbanUser(banner User, banned User) error {
+	_, err := db.c.Exec("DELETE FROM ban WHERE banner = ? AND banned = ?", banner.uId, banned.uId)
 	if err != nil {
 		// Errore nell'esecuzione della query
 		return err
@@ -23,8 +23,8 @@ func (db *appdbimpl) UnbanUser(banner integer, banned integer) error {
 }
 
 //GetBanList() ritorna la lista degli utenti bannati dalll'utente
-func (db *appdbimpl) GetBanList(user integer) ([]User,error) {
-	rows, err := db.c.Query("SELECT u.uId, u.name FROM user u, ban b WHERE u.uId = b.banned AND b.banner = ?  ", user)
+func (db *appdbimpl) GetBanList(user User) ([]User, error) {
+	rows, err := db.c.Query("SELECT u.uId, u.name FROM user u, ban b WHERE u.uId = b.banned AND b.banner = ?  ", user.uId)
 	if err != nil {
 		// Errore nell'esecuzione della query
 		return nil, err
@@ -33,7 +33,7 @@ func (db *appdbimpl) GetBanList(user integer) ([]User,error) {
 	// uso defer per ritardare l'operazione rows.close() fino alla fine della funzione
 	defer rows.Close()
 	// Array che conterra' tutti gli utenti bannati
-	var banList []User	// sono riempiti solo la variabile per il nome e per l'id!!!
+	var banList []User // sono riempiti solo la variabile per il nome e per l'id!!!
 
 	for rows.Next() {
 		var user User
@@ -45,7 +45,7 @@ func (db *appdbimpl) GetBanList(user integer) ([]User,error) {
 		banList = append(banList, user)
 	}
 
-	if rows.Err()!= nil {
+	if rows.Err() != nil {
 		// Errore
 		return nil, err
 	}
@@ -55,22 +55,20 @@ func (db *appdbimpl) GetBanList(user integer) ([]User,error) {
 }
 
 //IsBanned() controlla se A e' stato bannato da B
-func (db *appdbimpl) IsBanned(A integer,B integer) (bool,error) {
+func (db *appdbimpl) IsBanned(A User, B User) (bool, error) {
 
 	var result bool
-	err := db.c.QueryRow(`SELECT EXISTS ( SELECT 1 FROM ban WHERE banner = ? AND banned = ?) AS isBanned;`, B,A).Scan(&result)
+	err := db.c.QueryRow(`SELECT EXISTS ( SELECT 1 FROM ban WHERE banner = ? AND banned = ?) AS isBanned;`, B.uId, A.uId).Scan(&result)
 	if err != nil {
 		// Errore nell'esecuzione della query
-		return nil, err
+		return false, err
 	}
 
-	return result,nil
+	return result, nil
 }
 
-
-
-//GetBanList() ritorna la lista degli utenti bannati dalll'utente
-func (db *appdbimpl) GetBanListVINT(user integer) ([]integer,error) {
+//GetBanList() ritorna la lista degli utenti bannati dalll'utente, ma torna solo gli id
+func (db *appdbimpl) GetBanListVINT(user User) ([]int, error) {
 
 	rows, err := db.c.Query("SELECT banned FROM ban WHERE banner = ?", user)
 	if err != nil {
