+++ service/database/db-follow.go
@@ -1,37 +1,36 @@
 package database
 
-
 // lo stream dell'utente e' stato messo all'interno di db-user
 
 //  FollowUser() permette all'utente user di seguire userToFollow
-func (db *appdbimpl) FollowUser(user integer, userToFollow integer) error {		//A e' seguito da B
-	_, err := db.c.Exec("INSERT INTO follow (A,B) VALUES (?, ?)", userToFollow, user)
+func (db *appdbimpl) FollowUser(user User, userToFollow User) error { //A e' seguito da B
+	_, err := db.c.Exec("INSERT INTO follow (A,B) VALUES (?, ?)", userToFollow.uId, user.uId)
 
-	if err != nil{
+	if err != nil {
 		// Errore nell'esecuzione della Query
 		return err
 	}
-	//nessun Errore 
+	//nessun Errore
 	return nil
 }
+
 // UnfollowUser() permette all'utente di smettere di seguire followed
-func (db *appdbimpl) UnfollowUser(user integer, followed integer) error {	//A e' seguito da B
-	_, err := db.c.Exec("DELETE FROM follow WHERE A = ? AND B = ?", followed, user)
+func (db *appdbimpl) UnfollowUser(user User, followed User) error { //A e' seguito da B
+	_, err := db.c.Exec("DELETE FROM follow WHERE A = ? AND B = ?", followed.uId, user.uId)
 
-	if err != nil{
+	if err != nil {
 		// Errore nell'esecuzione della Query
 		return err
 	}
-	//nessun Errore 
+	//nessun Errore
 	return nil
 }
-
 
 //get my follower
 
-func (db *appdbimpl) GetFollowingList(user integer) ([]integer,error) {		//A e' seguito da B
+func (db *appdbimpl) GetFollowingList(userr User) ([]int, error) { //A e' seguito da B
 
-	rows, err := db.c.Query("SELECT A FROM follow WHERE B = ?  ", user)
+	rows, err := db.c.Query("SELECT A FROM follow WHERE B = ?  ", userr.uId)
 	if err != nil {
 		// Errore nell'esecuzione della query
 		return nil, err
@@ -40,19 +39,19 @@ func (db *appdbimpl) GetFollowingList(user integer) ([]integer,error) {		//A e' 
 	// uso defer per ritardare l'operazione rows.close() fino alla fine della funzione
 	defer rows.Close()
 
-	var following []integer	
+	var following []int
 
 	for rows.Next() {
-		var user integer
+		var user int
 		err := rows.Scan(&user)
 		if err != nil {
 			// Errore nella scansione della riga
 			return nil, err
 		}
-		following = append(banList, user)
+		following = append(following, user)
 	}
 
-	if rows.Err()!= nil {
+	if rows.Err() != nil {
 		// Errore
 		return nil, err
 	}
@@ -61,10 +60,9 @@ func (db *appdbimpl) GetFollowingList(user integer) ([]integer,error) {		//A e' 
 
 }
 
+func (db *appdbimpl) GetFollowerList(userr User) ([]int, error) { //A e' seguito da B
 
-func (db *appdbimpl) GetFollowerList(user integer) ([]integer,error) {		//A e' seguito da B
-
-	rows, err := db.c.Query("SELECT B FROM follow WHERE A = ?  ", user)
+	rows, err := db.c.Query("SELECT B FROM follow WHERE A = ?  ", userr.uId)
 	if err != nil {
 		// Errore nell'esecuzione della query
 		return nil, err
@@ -73,19 +71,19 @@ func (db *appdbimpl) GetFollowerList(user integer) ([]integer,error) {		//A e' s
 	// uso defer per ritardare l'operazione rows.close() fino alla fine della funzione
 	defer rows.Close()
 
-	var follower []integer	
+	var follower []int
 
 	for rows.Next() {
-		var user integer
+		var user int
 		err := rows.Scan(&user)
 		if err != nil {
 			// Errore nella scansione della riga
 			return nil, err
 		}
-		follower = append(banList, user)
+		follower = append(follower, user)
 	}
 
-	if rows.Err()!= nil {
+	if rows.Err() != nil {
 		// Errore
 		return nil, err
 	}
